requirements.txt
Flask>=2.0
Flask_SQLAlchemy>=2.5
dnspython>=2.0
